Remove decorators, improve error handling and timeouts
Decorators
----------
We have to remove decorators as they won't work in deno deploy's
playground, as they require a deno.json file which is not available in
single file setups

https://deno.com/deploy/changelog#es-decorators-are-enabled-on-deno-deploy-replacing-experimental-ts-decorators

Error handling
--------------
Add error logging on unexpected behaviour from speedrun.com or discord,
as well as responding to the user in case something unexpected happens.

Timeouts
--------
When the mention parameter is passed, mention the user telling them to
wait, then edit the message with the actual invite. When the parameter
is not passed, simply defer as before.